Revert change that breaks collector embedding
